Registration module updated

+++ sources/main/actions/registration/RegistrationLogin.java
@@ -43,7 +43,7 @@ public class RegistrationLogin extends HttpServlet {
 		PrintWriter pw=response.getWriter();
 		JSONObject data = new JSONObject();
 		HttpSession session = request.getSession();
-		 int reg_id=Integer.parseInt(request.getParameter("reg_id"));
+		Long reg_id=Long.parseLong(request.getParameter("reg_id"));
 		   int status=Query.retrieveRegistrationStatus(reg_id);
 		   System.out.println(status);
 		   if(status==1){
